requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
pycryptodome==3.23.0
requests==2.32.5
urllib3==2.5.0
